requirements.txt
chainer==1.16.0
click==6.6
dask==0.11.1
decorator==4.0.10
dlib==19.1.0
filelock==2.0.6
Flask==0.11.1
Flask-SocketIO==2.7.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
msgpack-python==0.4.8
networkx==1.11
nose==1.3.7
numpy==1.11.2
Pillow==3.4.1
protobuf==3.0.0
PyOpenGL==3.1.0
PyOpenGL-Demo==3.0.0
python-engineio==1.0.3
python-socketio==1.6.0
scikit-image==0.12.3
scipy==0.18.1
six==1.10.0
tensorflow==0.11.0rc0
toolz==0.8.0
Werkzeug==0.11.11
